I'm the one who wrote the phone number on the card
I'm sorry that I'm using this strange way to get communication with you guys, remember that phone number on the card? Yes, that's me, I know that's crazy, and I don't know why I did that. Maybe it's because I like the owner of that hat a lot and I really want to maybe just be a friends of her, hope it'll not get myself in to trouble...😅
and..... if you guys interested in some drone videos of Lynn, Boston, or somewhere else of MA, check my YouTube channel https://www.youtube.com/watch?v=txpnFjsL3NI&t=47s
this is the newest one, hope you guys enjoyed it(and I know all these are so strange...)(I'm a strange person)
if you want to say something to me, here's my email 2521355956@qq.com